+++ source/CHANGELOG.md
@@ -3,9 +3,11 @@
 
 Development
 -------------
+NEW:
+* New analysis function AnnualRecurrenceProbability which calculates the annual recurrence probility and return period of annual maxima using plotting position method
+
 API-CHANGES:
-* New property `TimeSeries.MaximumNode()` for retrieving the node (date and value) with the maximum value of a time series, 
-  optionally within a defined time period
+* New property `TimeSeries.MaximumNode()` for retrieving the node (date and value) with the maximum value of a time series, optionally within a defined time period
 
 Version 2.6.1
 -------------